Multiple Fixes to succesfully compile on Ubuntu, Gentoo and Mac OS (Big Sur)

+++ libs/LzmaLib/source/MtCoder.h
@@ -107,7 +107,7 @@ typedef struct _CMtCoder
     SRes writeRes;
     unsigned writeIndex;
     Byte ReadyBlocks[MTCODER__BLOCKS_MAX];
-    LONG numFinishedThreads;
+    Int32 numFinishedThreads;
   #endif
 
   unsigned numStartedThreadsLimit;
